requirements-dev.txt
pytest
pytest-mock
pytest-cov
mock
svtools.ipypi_testutils
